backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-countries==7.6.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
phonenumbers==8.13.50
pluggy==1.5.0
PyJWT==2.10.0
pytest==8.3.3
pytest-django==4.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
